SMD-842 - Moving table validation under 'core/validation'
Although table validation for Pathfinders is strongly linked to the table extraction process (they are configured within the same schema, and share the use of the Table class), I feel that it should be moved to the 'core/validation' directory, as that is where you would expect to find all validation logic, not just for TF validation logic. Moving the validation-specific modules 'core/table_extraction/checks.py' and 'core/table_extraction/validate.py' is not the end of the story, as we then need to adjust 'core/table_extraction/__init__.py' to remove the imports from those modules, as it does not make sense for an '__init__.py' file to be importing from a completely different module. Further, 'core/table_extraction/exceptions.py' needs to be broken in two, to allow the validation-specific 'TableValidationError' and 'TableValidationErrors' to move alongside the 'checks' and 'validate' modules into 'core/validation'. Then, the full pipeline test needs to be moved one level up the test directory hierarchy, as it now concerns not only code in 'core/table_extraction' but also code in 'core/validation'; the E2E CSV resources likewise need to move, and given that we are creating a new 'tests/resources/pathfinders' folder, it makes sense also to move 'tests/resources/extracted_data.py' into it while we're at it. Finally, we needed to adjust the basename of the migrating 'test_validate.py' file, to avoid a conflict error when tests are run.
