+++ core/table_extraction/__init__.py
@@ -1,18 +1,11 @@
-import core.table_extraction.checks  # noqa - adds custom checks to panderas built in set of checks
-from core.table_extraction.exceptions import (
-    TableExtractionError,
-    TableProcessingError,
-    TableValidationErrors,
-)
+import core.validation.pathfinders.schema_validation.checks  # noqa - adds custom checks to panderas built in set of checks
+from core.table_extraction.exceptions import TableExtractionError, TableProcessingError
 from core.table_extraction.extract import TableExtractor
 from core.table_extraction.process import TableProcessor
-from core.table_extraction.validate import TableValidator
 
 __all__ = [
     "TableExtractor",
     "TableProcessor",
-    "TableValidator",
-    "TableProcessingError",
     "TableExtractionError",
-    "TableValidationErrors",
+    "TableProcessingError",
 ]
