+++ core/table_extraction/exceptions.py
@@ -1,6 +1,3 @@
-from core.table_extraction.table import Cell
-
-
 class TableExtractionError(Exception):
     def __init__(self, message: str):
         super().__init__(message)
@@ -9,21 +6,3 @@ class TableExtractionError(Exception):
 class TableProcessingError(Exception):
     def __init__(self, message: str):
         super().__init__(message)
-
-
-class TableValidationError:
-    """Validation error with reference to an Excel cell."""
-
-    message: str
-    cell: Cell | None
-
-    def __init__(self, message: str, cell: Cell | None):
-        self.message = message
-        self.cell = cell
-
-
-class TableValidationErrors(Exception):
-    validation_errors: list[TableValidationError]
-
-    def __init__(self, validation_errors: list[TableValidationError]):
-        self.validation_errors = validation_errors
