+++ core/validation/pathfinders/schema_validation/checks.py
