+++ tests/resources/pathfinders/extracted_data.py
