+++ tests/validation_tests/pathfinders/schema_validation_tests/test_checks.py
@@ -3,8 +3,8 @@ from functools import partial
 
 import pytest
 
-from core.table_extraction.checks import max_word_count, not_in_future, postcode_list
 from core.table_extraction.config.pf_r1_config import PFRegex
+from core.validation.pathfinders.schema_validation.checks import max_word_count, not_in_future, postcode_list
 
 
 @pytest.mark.parametrize(
